+++ src/upnp_desc_handler.h
@@ -36,7 +36,7 @@ class UpnpDescHandler : public RequestHandler {
 public:
     explicit UpnpDescHandler(const std::shared_ptr<Content>& content, const std::shared_ptr<UpnpXMLBuilder>& xmlBuilder);
 
-    const struct ClientInfo* getInfo(const char* filename, UpnpFileInfo* info) override;
+    const struct ClientObservation* getInfo(const char* filename, UpnpFileInfo* info) override;
     std::unique_ptr<IOHandler> open(const char* filename, const std::shared_ptr<Quirks>& quirks, enum UpnpOpenFileMode mode) override;
 
 private: