+++ ChangeLog.md
@@ -13,6 +13,7 @@
 - Refactor Upnp Services
 - Upgrade cxxopts.hpp to 3.2.1
 - Upgrade md5 to 2002 version
+- Use full client info as requestCookie
 - Warn about empty script result
 - Web security and upnp compatibility